Code refactor and add tests
