+++ event-bus/nuls-event-bus/src/main/java/io/nuls/eventbus/rpc/processor/SendRetryProcessor.java
@@ -20,6 +20,7 @@ public class SendRetryProcessor implements Runnable {
     SendRetryProcessor(Object[] obj){
         this.subscriberEvent = obj;
     }
+
     @Override
     public void run() {
         try{
