+++ event-bus/nuls-event-bus/src/test/java/io/nuls/eventbus/test/service/EbStorageServiceTest.java
@@ -7,6 +7,7 @@ import io.nuls.eventbus.model.Topic;
 import io.nuls.eventbus.service.EbStorageService;
 import io.nuls.eventbus.service.EbStorageServiceImpl;
 import io.nuls.rpc.model.ModuleE;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -46,11 +47,13 @@ public class EbStorageServiceTest {
         for(Subscriber sub : fromStorage.getSubscribers()){
             System.out.println("Subscriber module:"+sub.getModuleAbbr());
         }
+        Assert.assertTrue("ac_create".equals(fromStorage.getTopicId()));
     }
 
     @Test
     public void loadAllTopics(){
         ConcurrentMap<String,Topic> topicMap = ebStorageService.loadTopics();
         System.out.println(topicMap.keySet().toString());
+        Assert.assertNotNull(topicMap);
     }
 }
